+++ packages/terra-application-utility/src/utility/_UtilityMenu.jsx
@@ -42,9 +42,17 @@ const propTypes = {
    */
   onRequestClose: PropTypes.func,
   /**
+   * Role attribute of Utility Menu.
+   */
+  role: PropTypes.string,
+  /**
    * Sets the Utility variant. One of Utils.VARIANTS.HEADER, Utils.VARIANTS.MENU.
    */
   variant: PropTypes.oneOf([Utils.VARIANTS.HEADER, Utils.VARIANTS.MENU]),
+};
+
+const defaultProps = {
+  role: 'navigation',
 };
 
 const processMenuItems = (items) => {
@@ -246,6 +254,7 @@ class UtilityMenu extends React.Component {
       onChange,
       onRequestClose,
       variant,
+      role,
       ...customProps
     } = this.props;
 
@@ -367,7 +376,7 @@ class UtilityMenu extends React.Component {
           footer={footer}
           fill={isHeightBounded}
           className={menuClassNames}
-          role="navigation"
+          role={role}
           aria-label={menuText}
         >
           {this.buildListContent(currentItem)}
@@ -381,4 +390,5 @@ class UtilityMenu extends React.Component {
 UtilityMenu.propTypes = propTypes;
 UtilityMenu.processMenuItems = processMenuItems;
 UtilityMenu.hasChevron = hasChevron;
+UtilityMenu.defaultProps = defaultProps;
 export default injectIntl(UtilityMenu);